typing fixes

+++ src/Bot/Events/messageEvents/messageDelete/giveaway.ts
@@ -2,11 +2,13 @@ import type DDeno from 'discordeno';
 import client from '../../../BaseClient/DDenoClient.js';
 
 export default async (msg: DDeno.Message) => {
+  if (!msg.guildId) return;
+
   client.ch.query(`DELETE FROM giveaways WHERE msgid = $1;`, [String(msg.id)]);
 
-  const g = client.giveaways.get(`${msg.id}-${msg.guildId}`);
-  if (g) {
-    g.cancel();
-    client.giveaways.delete(`${msg.id}-${msg.guildId}`);
-  }
+  const guildGiveawayCache = client.giveaways.get(msg.guildId);
+  guildGiveawayCache?.get(msg.id)?.cancel();
+
+  if (guildGiveawayCache?.size === 1) client.giveaways.delete(msg.guildId);
+  else guildGiveawayCache?.delete(msg.id);
 };
